+++ API/Controllers/UsersController.cs
@@ -1,8 +1,11 @@
-﻿using API.Data;
+﻿using System.Security.Claims;
+using API.Data;
 using API.Interfaces;
 using DomainModels.Dto.UserDto;
+using DomainModels.Enums;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace API.Controllers;
 
@@ -15,16 +18,18 @@ public class UsersController : ControllerBase
 {
 	private readonly IUserService _userService;
 	private readonly ILogger<UsersController> _logger;
+	private readonly AppDBContext _context;
 
 	/// <summary>
 	/// Initializes a new instance of the <see cref="UsersController"/> class.
 	/// </summary>
 	/// <param name="userService">The user service for user operations.</param>
 	/// <param name="logger">The logger instance for logging errors and information.</param>
-	public UsersController(IUserService userService, ILogger<UsersController> logger)
+	public UsersController(IUserService userService, ILogger<UsersController> logger, AppDBContext context)
 	{
 		_userService = userService;
 		_logger = logger;
+		_context = context;
 	}
 
 	/// <summary>
@@ -86,6 +91,43 @@ public class UsersController : ControllerBase
 			_logger.LogError(ex, "Error occurred while retrieving user with ID {Id}.", id);
 			return StatusCode(500, "An unexpected error occurred.");
 		}
+	}
+	/// <summary>
+	/// Retrieves the currently authenticated user's profile information.
+	/// </summary>
+	/// <returns>
+	/// An object containing user ID, email, username, creation date, last login, and role if found;
+	/// <see cref="UnauthorizedObjectResult"/> if the user ID is not found in the token;
+	/// <see cref="NotFoundObjectResult"/> if the user does not exist in the database.
+	/// </returns>
+	[Authorize]
+	[HttpGet("me")]
+	public IActionResult GetCurrentUser()
+	{
+		var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+		if (userId == null)
+			return Unauthorized("User-ID was not found in a token.");
+
+		var user = _context.Users
+			.Include(u => u.UserRole)
+			.FirstOrDefault(u => u.Id.ToString() == userId);
+
+		if (user == null)
+			return NotFound("User was not found in a database.");
+
+		var roleEnum = (RoleEnum)user.UserRoleId;
+
+		return Ok(new
+		{
+			Id = user.Id,
+			Email = user.Email,
+			Username = user.UserName,
+			CreatedAt = user.CreatedAt,
+			LastLogin = user.LastLogin,
+			Role = user.UserRole.RoleName.ToString(),
+			Description = roleEnum.GetDescription()
+		});
 	}
 
 	/// <summary>
